requirements.txt
coverage==5.5
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.4
monai==0.4.0
networkx==2.5
nibabel==3.2.1
numpy==1.20.0
packaging==20.9
pandas==1.2.1
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.1
scipy==1.6.0
SimpleITK==2.0.2
six==1.15.0
tifffile==2021.1.14
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
typing-extensions==3.7.4.3
